% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_mean.R
\name{interpolate_mean}
\alias{interpolate_mean}
\title{Interpolate Mean Accumulated Values for Spatio-Temporal Hawkes Process}
\usage{
interpolate_mean(
  results,
  data,
  Sites1,
  Sites2,
  Xw,
  Xm,
  Xu,
  Xmr,
  Xwr,
  Xur,
  R3,
  tau,
  d
)
}
\arguments{
\item{results}{The output from the \code{STModelHawkesMCMC} function, containing posterior samples of model parameters.}

\item{data}{A matrix of event times for the phenomenon of interest, where rows represent time points and columns represent locations.}

\item{Sites1}{A matrix of locations used to estimate the parameters of the spatio-temporal Hawkes process.}

\item{Sites2}{A vector representing the location where the mean accumulated value will be interpolated.}

\item{Xw}{A matrix of covariates for \(W\) at the locations in \code{Sites1}.}

\item{Xm}{A matrix of covariates for \(M\) at the locations in \code{Sites1}.}

\item{Xu}{A matrix of covariates for \(U\) at the locations in \code{Sites1}.}

\item{Xmr}{A matrix of covariates for \(M\) at the interpolation location (\code{Sites2}).}

\item{Xwr}{A matrix of covariates for \(W\) at the interpolation location (\code{Sites2}).}

\item{Xur}{A matrix of covariates for \(U\) at the interpolation location (\code{Sites2}).}

\item{R3}{A scalar radius defining the neighborhood for interpolation.}

\item{tau}{A vector of times where the mean accumulated values will be interpolated.}

\item{d}{A positive real number used in the distance weighting function.}
}
\value{
A matrix (\code{MatMean}) of interpolated mean accumulated values, where rows correspond to posterior samples, and columns correspond to time points in \code{tau}.
}
\description{
This function interpolates the mean accumulated values for a spatio-temporal Hawkes process
using the posterior samples from the \code{STModelHawkesMCMC} function, spatial and temporal data,
and user-defined covariates.
}
\details{
This function combines spatial and temporal information to compute mean accumulated values at a specified location and time points. The computation is based on posterior samples of the Hawkes process parameters \eqn{\gamma}, \eqn{\alpha}, \eqn{\eta}, \eqn{\beta}, with spatial interpolation using a distance-based weighting function.
}
