% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STModelHawkesMCMC.R
\name{STModelHawkesMCMC}
\alias{STModelHawkesMCMC}
\title{Estimate Parameters of Spatio-Temporal Hawkes Model using MCMC}
\usage{
STModelHawkesMCMC(
  data,
  sites,
  Xw = as.matrix(cbind(as.matrix(rep(1, n)), sites)),
  Xm = Xw,
  Xu = Xw,
  prior = list(Nu = 2, Aw = as.matrix(rep(0, ncol(Xw))), Bw = diag(100, ncol(Xw)), Am =
    as.matrix(rep(0, ncol(Xm))), Bm = diag(100, ncol(Xm)), Au = as.matrix(rep(0,
    ncol(Xu))), Bu = diag(100, ncol(Xu)), Aa = 2, aaw = 0.1, bbw = 0.1, aam = 1, bbm = 1,
    aau = 1, bbu = 1, aa1w = 0.1, bb1w = 0.1, aa1m = 0.1, bb1m = 0.1, aa1u = 0.1, bb1u =
    0.1),
  iteration,
  burnin,
  jump
)
}
\arguments{
\item{data}{A vector of event times for the phenomenon of interest. This input is similar to the \code{data} argument in the \code{STModelWeibullMCMC} function.}

\item{sites}{A matrix of geographic coordinates corresponding to the event locations. This input is similar to the \code{sites} argument in the \code{STModelWeibullMCMC} function.}

\item{Xw}{A matrix of covariates associated with the \(W\) process. Defaults to a matrix combining an intercept and \code{sites}.}

\item{Xm}{A matrix of covariates associated with the \(M\) process. Defaults to the same as \code{Xw}.}

\item{Xu}{A matrix of covariates associated with the \(U\) process. Defaults to the same as \code{Xw}.}

\item{prior}{A list of hyperparameters defining the prior distributions in the model.}

\item{iteration}{Number of MCMC iterations. Defaults to 1000.}

\item{burnin}{Number of iterations for the burn-in period. Defaults to 900.}

\item{jump}{Thinning parameter for the MCMC sampling. Defaults to 10.}
}
\value{
A list containing:
\itemize{
\item Estimated parameters for the spatio-temporal Hawkes model.
\item MCMC samples for each parameter.
\item Convergence diagnostics and summary statistics.
}
}
\description{
This function estimates the parameters of a spatio-temporal Hawkes model
using Markov Chain Monte Carlo (MCMC) methods.
}
